Improve usability and documentation of imgproc.
The usability is mainly improved by adding command line arguments for
input and output files of the various main functions.

The documentation is improved by adding a readme file which explains the
image processing part required to generate a trajectory as well as
explanation of the various runnable main-functions in the imgproc side of
the project.

Additionally, the python logo was added as an example image, that the
various examples are tuned after.